+++ crates/hooks/src/lib.rs
@@ -15,7 +15,6 @@ mod use_focus;
 mod use_node;
 mod use_platform;
 mod use_theme;
-mod use_ticker;
 
 #[cfg(feature = "use_camera")]
 mod use_camera;
@@ -34,7 +33,6 @@ pub use use_focus::*;
 pub use use_node::*;
 pub use use_platform::*;
 pub use use_theme::*;
-pub use use_ticker::*;
 
 #[cfg(feature = "use_camera")]
 pub use use_camera::*;
